refactor(codegen): split escape_news collector into topical files (#1296)
`crates/perry-codegen/src/collectors/escape_news.rs` was 2034 lines and
failed the 2,000-line file-size CI gate. Splitting it into topical sibling
modules under `collectors/` — no behavior change; function bodies move
verbatim and the same public/internal API is re-exported from
`collectors/mod.rs`.

Moves:
- `this_as_value.rs` — `class_uses_this_as_value`,
  `class_chain_extends_builtin_error`, `stmts_use_this_as_value`,
  `expr_uses_this_as_value`.
- `class_accessors.rs` — `is_class_getter`, `is_class_setter`.
- `escape_check.rs` — `find_new_candidates`, `check_escapes_in_stmts`,
  `check_escapes_in_expr`.
- `local_refs.rs` — `expr_contains_local_get`,
  `mark_all_candidate_refs_in_expr`.

`escape_news.rs` keeps the escape-collection entry points
(`collect_non_escaping_news`, `collect_non_escaping_new_used_fields` and
the two private helpers) plus the trailing `MAX_SCALAR_ARRAY_LEN` const.

Final line counts: escape_news 693, escape_check 870, this_as_value 337,
local_refs 114, class_accessors 50 — all comfortably under the 2,000 gate.
