+++ crates/perry-codegen/src/collectors/class_accessors.rs
@@ -1,0 +1,50 @@
+//! Getter / setter dispatch detection used by escape analysis.
+//!
+//! Split out of `escape_news.rs` in v0.5.1021 to satisfy the file-size CI
+//! gate. No behavior change — these functions remain `pub` and are re-
+//! exported from `collectors/mod.rs`.
+
+/// Is `property` a getter on `class_name` (walking its inheritance chain)?
+/// Used by escape analysis: a `LocalGet(candidate).gettableProp` access is
+/// a real getter dispatch that needs `this` as a heap pointer, so the
+/// candidate must escape.
+pub fn is_class_getter(
+    classes: &std::collections::HashMap<String, &perry_hir::Class>,
+    class_name: &str,
+    property: &str,
+) -> bool {
+    let mut cur = Some(class_name.to_string());
+    while let Some(name) = cur {
+        if let Some(class) = classes.get(&name) {
+            if class.getters.iter().any(|(n, _)| n == property) {
+                return true;
+            }
+            cur = class.extends_name.clone();
+        } else {
+            return false;
+        }
+    }
+    false
+}
+
+/// Mirror of `is_class_getter` for setters — used on the PropertySet/
+/// PropertyUpdate paths where a setter dispatch (vs. a plain field write)
+/// likewise needs a real `this` pointer.
+pub fn is_class_setter(
+    classes: &std::collections::HashMap<String, &perry_hir::Class>,
+    class_name: &str,
+    property: &str,
+) -> bool {
+    let mut cur = Some(class_name.to_string());
+    while let Some(name) = cur {
+        if let Some(class) = classes.get(&name) {
+            if class.setters.iter().any(|(n, _)| n == property) {
+                return true;
+            }
+            cur = class.extends_name.clone();
+        } else {
+            return false;
+        }
+    }
+    false
+}
