+++ crates/perry-codegen/src/collectors/mod.rs
@@ -6,18 +6,22 @@
 //! hub — public-API shape (`crate::collectors::*`) is preserved.
 
 mod clamp_detect;
+mod class_accessors;
 mod closures;
 mod escape_arrays;
+mod escape_check;
 mod escape_news;
 mod escape_objects;
 mod i32_locals;
 mod i64_emit;
 mod index_uses;
 mod integer_locals;
+mod local_refs;
 mod mutation;
 mod pointer_locals;
 mod refs;
 mod shadow_slots;
+mod this_as_value;
 
 // Public re-exports for the visible API (`pub fn emit_i64_function` etc.).
 pub use clamp_detect::{detect_clamp3, detect_clamp_u8, is_integer_specializable, returns_integer};
@@ -26,17 +30,15 @@ pub use i64_emit::emit_i64_function;
 // Internal-to-crate re-exports — explicit names because globs don't
 // transitively expose through `pub(crate) use crate::collectors::*`.
 pub(crate) use clamp_detect::{i64s_expr, i64s_stmts, returns_int_expr, returns_int_stmts};
+pub(crate) use class_accessors::{is_class_getter, is_class_setter};
 pub(crate) use closures::{collect_closures_in_expr, collect_closures_in_stmts};
 pub(crate) use escape_arrays::{
     check_array_escapes_in_expr, check_array_escapes_in_stmts, collect_non_escaping_arrays,
     const_index, find_array_candidates, MAX_SCALAR_OBJECT_FIELDS,
 };
+pub(crate) use escape_check::{check_escapes_in_expr, check_escapes_in_stmts, find_new_candidates};
 pub(crate) use escape_news::{
-    check_escapes_in_expr, check_escapes_in_stmts, class_chain_extends_builtin_error,
-    class_uses_this_as_value, collect_non_escaping_new_used_fields, collect_non_escaping_news,
-    expr_contains_local_get, expr_uses_this_as_value, find_new_candidates, is_class_getter,
-    is_class_setter, mark_all_candidate_refs_in_expr, stmts_use_this_as_value,
-    MAX_SCALAR_ARRAY_LEN,
+    collect_non_escaping_new_used_fields, collect_non_escaping_news, MAX_SCALAR_ARRAY_LEN,
 };
 pub(crate) use escape_objects::{
     check_object_literal_escapes_in_expr, check_object_literal_escapes_in_stmts,
@@ -59,6 +61,7 @@ pub(crate) use integer_locals::{
     collect_extra_integer_let_ids, collect_flat_row_aliases, collect_integer_locals,
     is_int32_producing_expr,
 };
+pub(crate) use local_refs::{expr_contains_local_get, mark_all_candidate_refs_in_expr};
 pub(crate) use mutation::{expr_has_mutation, has_any_mutation, is_local_get_chain};
 pub(crate) use pointer_locals::collect_pointer_typed_locals;
 pub(crate) use refs::{
@@ -67,4 +70,8 @@ pub(crate) use refs::{
 pub(crate) use shadow_slots::{
     collect_declared_shadow_locals_in_stmt, collect_declared_shadow_slots_in_stmts,
     collect_shadow_slot_clear_points,
+};
+pub(crate) use this_as_value::{
+    class_chain_extends_builtin_error, class_uses_this_as_value, expr_uses_this_as_value,
+    stmts_use_this_as_value,
 };
